Add as a maven repo
