--- src/overview.html
@@ -1,35 +1,0 @@
-This library is free software; you can redistribute it and/or
-modify it under the terms of the GNU Lesser General Public
-License as published by the Free Software Foundation; either
-version 2.1 of the License, or (at your option) any later version.
-
-This library is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public
-License along with this library; if not, write to the Free Software
-Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
-MA 02110-1301  USA
--->
-<!DOCTYPE xhtml>
-<html>
-    <head>
-        <title>JMBDiscId Java libdiscid interface</title>
-        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
-    </head>
-    <body>
-        <div>This library enables Java programms to calculate DiscIds for <a href="http://musicbrainz.org">MusicBrainz.org</a>
-            and <a href="http://freedb.org">FreeDB.org</a> and generate a submittion URL and a data lookup url for <a href="http://musicbrainz.org">MusicBrainz.org</a>
-        </div>
-        <div>
-            An example is provided in the examples package; Have a look there for reference.
-        </div>
-        <div>
-            to run the example just call: java JMBDiscId.jar &lt;path to lib including lib name and extention&gt; &lt;drive path to use&gt;
-        </div>
-        <div>My cmd-line looks like this: java -jar "V:\Programming\Netbeans\JMBDiscId\dist\JMBDiscId.jar" "V:\Programming\Netbeans\JMBDiscId\libdiscid\discid.dll" "H:"
-        </div>
-    </body>
-</html>
